Add more consistency tests for QualityUbSolver
