+++ raphael-solver/src/step_lower_bound_solver/tests.rs
@@ -1,57 +1,16 @@
 use raphael_sim::*;
-use test_case::test_matrix;
 
 use crate::{
     AtomicFlag, SolverSettings,
     actions::{FULL_SEARCH_ACTIONS, use_action_combo},
+    test_utils::*,
 };
 
 use super::*;
 
-fn random_effects(settings: &Settings, rng: &mut impl rand::Rng) -> Effects {
-    let mut effects = Effects::new()
-        .with_inner_quiet(rng.random_range(0..=10))
-        .with_great_strides(rng.random_range(0..=3))
-        .with_innovation(rng.random_range(0..=4))
-        .with_veneration(rng.random_range(0..=4))
-        .with_waste_not(rng.random_range(0..=8))
-        .with_adversarial_guard(rng.random() && settings.adversarial)
-        .with_allow_quality_actions(rng.random() || !settings.backload_progress);
-    if settings.is_action_allowed::<Manipulation>() {
-        effects.set_manipulation(rng.random_range(0..=8));
-    }
-    if settings.is_action_allowed::<TrainedPerfection>() {
-        effects.set_trained_perfection_available(rng.random());
-        effects
-            .set_trained_perfection_active(!effects.trained_perfection_available() && rng.random());
-    }
-    if settings.is_action_allowed::<HeartAndSoul>() {
-        effects.set_heart_and_soul_available(rng.random());
-        effects.set_heart_and_soul_active(!effects.heart_and_soul_available() && rng.random());
-    }
-    if settings.is_action_allowed::<QuickInnovation>() {
-        effects.set_quick_innovation_available(rng.random());
-    }
-    effects
-}
-
-fn random_state(settings: &SolverSettings, rng: &mut impl rand::Rng) -> SimulationState {
-    SimulationState {
-        cp: rng.random_range(0..=settings.max_cp()),
-        durability: rng
-            .random_range(1..=settings.max_durability())
-            .next_multiple_of(5),
-        progress: rng.random_range(0..settings.max_progress()),
-        quality: rng.random_range(0..=settings.max_quality()),
-        unreliable_quality: 0,
-        effects: random_effects(&settings.simulator_settings, rng),
-    }
-    .try_into()
-    .unwrap()
-}
-
-/// Test that the StepLbSolver is consistent.
+/// Test that the StepLbSolver is consistent and admissible.
 /// It is consistent if the step-lb of a parent state is never greater than the step-lb of a child state.
+/// It is admissible if the step-lb of a state is never greater than the step count of a reachable final state.
 fn check_consistency(solver_settings: SolverSettings) {
     let mut solver = StepLbSolver::new(solver_settings, AtomicFlag::default());
     let mut rng = rand::rng();
@@ -89,7 +48,7 @@ const WITH_SPECIALIST_ACTIONS: ActionMask = REGULAR_ACTIONS
     .add(Action::HeartAndSoul)
     .add(Action::QuickInnovation);
 
-#[test_matrix(
+#[test_case::test_matrix(
     [20, 35, 60, 80],
     [REGULAR_ACTIONS, NO_MANIPULATION, WITH_SPECIALIST_ACTIONS]
 )]
